# Remove usage of regexp