--- a/.golangci.yml
+++ b/.golangci.yml
@@ -46,8 +46,6 @@ linters:
               desc: Use github.com/stretchr/testify/require instead of github.com/stretchr/testify/assert
             - pkg: io/ioutil
               desc: Use corresponding 'os' or 'io' functions instead.
-            - pkg: regexp
-              desc: Use github.com/grafana/regexp instead of regexp
             - pkg: github.com/pkg/errors
               desc: Use 'errors' or 'fmt' instead of github.com/pkg/errors
             - pkg: golang.org/x/exp/slices