--- a/go.mod
+++ b/go.mod
@@ -3,5 +3,3 @@ module github.com/prometheus/otlptranslator
 go 1.23.0
 
 toolchain go1.24.1
-
-require github.com/grafana/regexp v0.0.0-20240518133315-a468a5bfb3bc